+++ src/aideas/web.py
@@ -3,13 +3,30 @@ from flask import Flask, render_template, request
 import uuid
 import logging.config
 
+from aideas.app.request_data import RequestData, ValidationError
 from app.app import App
 from app.env import get_app_port, is_production
 from app.task import get_task, stop_task, TaskError
-from app.web_service import WebService, ValidationError
+from app.web_service import WebService
 
 web_app = Flask(__name__)
 # CORS(web_app)
+
+
+INDEX_TEMPLATE = 'index.html'
+AUTOMATION_INDEX_TEMPLATE = 'automate/index.html'
+
+
+@web_app.errorhandler(ValidationError)
+def handle_validation_error(e):
+    return render_template(
+        AUTOMATION_INDEX_TEMPLATE, **web_service.automation_index({"error": e.message})), 400
+
+
+@web_app.errorhandler(TaskError)
+def handle_validation_error(e):
+    return render_template(
+        AUTOMATION_INDEX_TEMPLATE, **web_service.automation_index({"error": e.message})), 400
 
 
 @web_app.route('/')
@@ -17,20 +34,22 @@ def index():
     return render_template('index.html', **web_service.index())
 
 
-AUTOMATION_INDEX_TEMPLATE = 'automate/index.html'
-
-
 @web_app.route('/' + AUTOMATION_INDEX_TEMPLATE)
 def automaton_index():
     return render_template(AUTOMATION_INDEX_TEMPLATE, **web_service.automation_index())
 
 
-@web_app.route('/automate/details/<tag>.html')
-def automation_details_form(tag: str):
-    if not tag:
-        error = {"error": "Specify what automation using a 'tag'."}
-        render_template(AUTOMATION_INDEX_TEMPLATE, {**web_service.automation_index(), **error})
-    return render_template(f'automate/details/{tag}.html', **web_service.automation_details_form(tag))
+@web_app.route('/automate/select-agents.html')
+def select_automaton_agents():
+    return render_template('automate/select-agents.html',
+                           **web_service.select_automation_agents(RequestData.require_tag(request)))
+
+
+@web_app.route('/automate/enter-details.html')
+def enter_automation_details():
+    data: dict[str, any] = RequestData.automation_details(request)
+    return render_template(
+        'automate/enter-details.html', **web_service.enter_automation_details(data))
 
 
 TASK_INDEX_TEMPLATE = 'task/index.html'
@@ -38,20 +57,18 @@ TASK_INDEX_TEMPLATE = 'task/index.html'
 
 @web_app.route('/automate/start', methods=['POST'])
 def start_automation():
-    asynch = request.args.get('async', True)
     task_id = str(uuid.uuid4().hex)
 
-    if asynch is True:
-        web_service.start_automation_async(task_id, request.form, request.files)
+    data = RequestData.task_config(task_id, request)
+
+    if data['async'] is True:
+        web_service.start_automation_async(task_id, data)
         return _render_task_index_template(f"Submitted task: {task_id}")
 
-    try:
-        task = web_service.start_automation(task_id, request.form, request.files)
-        args = {"info": f'<span style="font-size:0.75rem;">{task.to_html()}</span>'}
-    except Exception as ex:
-        args = _convert_to_error_obj(ex)
+    task = web_service.start_automation(task_id, data)
+    args = {"info": f'<span style="font-size:0.75rem;">{task.to_html()}</span>'}
     template_args = {**web_service.automation_index(), **args}
-    return render_template('automate/index.html', **template_args)
+    return render_template(AUTOMATION_INDEX_TEMPLATE, **template_args)
 
 
 @web_app.route('/' + TASK_INDEX_TEMPLATE)
@@ -91,14 +108,6 @@ def _get_task_links(task_id: str) -> dict[str, any]:
 
 def _render_task_index_template(info: str = None):
     return render_template(TASK_INDEX_TEMPLATE, **web_service.tasks(_get_task_links, info))
-
-
-def _convert_to_error_obj(ex: Exception) -> dict[str, str]:
-    default_err_msg = "An unexpected error occurred while trying to automate."
-    if isinstance(ex, ValidationError) or isinstance(ex, TaskError):
-        return {"error": ex.message if ex.message else default_err_msg}
-    else:
-        return {"error": default_err_msg}
 
 
 if __name__ == '__main__':